Switch browser scroll to rec. panel when read more visible
Makes the main browser scrollbar and interactions (e.g. keyboard, mouse wheel scrolling) apply to the recommendations panel when read-more is visible. To do this while preserving the user's previous scroll position, we need to apply the previous scroll as a negative offset to that scrollable section while setting it position: fixed. This requires a wrapper element. <section> elements are used to clearly distinguish the two scrollable sections.