Rec tabs merged (#85)
* Distinguish detected and undetected tabs (#63)

Visually seperates tabs for selecting detected vs not detected issues. Adds a button to show/hide undetected tabs, and slightly refactors open/close logic to reduce repetition and move recommendations logic to recommendations.js. Adds lead-in text to undetected issues articles to explicitly explain that they're issues that haven't been detected, and slightly rewords other messages to match.

* Select default tab if selected undetected tab is hidden

If the user selects an article about an undetected issue then closes the undetected tabs panel, switch the selection away from the tab in the now-hidden panel to the default detected tab

* Include pseudoelement content in alert width measurement

With the non-category-specific part of the alert message defined like other boilerplate in CSS ::before or ::after content, we need to include pseudoelements when measuring the alert's width.

* Single chevrons on show / hide

* Drop multi-column layout

* Improve paragraph readability

* Limit expanded read more width

* Pregenerate read more menu based on content headings

* Use better naming and apply styling to article menu

* Remove columns div, move its CSS and menu to its parent

* Freeze body scroll when opening scrollable article

* Keep track of article headings when reapplying the template

* Keep article menu items always in sync with article headings

* Highlight current article section in the article menu

* Smooth animated scrolling on jump links

* Fix typo

* Replace kebabified section ids with sequential ones

* Fix article menu not sticking when screen height <700px

* Switch browser scroll to rec. panel when read more visible

Makes the main browser scrollbar and interactions (e.g. keyboard, mouse wheel scrolling) apply to the recommendations panel when read-more is visible. To do this while preserving the user's previous scroll position, we need to apply the previous scroll as a negative offset to that scrollable section while setting it position: fixed. This requires a wrapper element. <section> elements are used to clearly distinguish the two scrollable sections.

* Remove obsolete comment

* add appveyor badge, fix npm link and license link

* more badge fixes

* fix npm package link for scoped packages

* scoped packages require www apparently

* Use heavier angle bracket characters for show/hide

* Revert "Switch browser scroll to rec. panel when read more visible"

This reverts commit 652186b8d8ae6e24fff5f379121d2fac07c70ab4

* Fix position: sticky; on nav on all screen heights

* Linting fixes

* Fixing small bugs left by merge

* Highlight nav link on returning to viewed tabs

Fixes a small bug where the top nav link isn't highlighted if a user returns to a tab they had previously read

* If width <670px allow article menu to move above

* Fix CSS typo

* Remove obsolete CSS