Use HMM to seperate node from v8 in cpu  (#149)
* Revert "try to filter cpu noise in issueCategory (#134)"

This reverts commit 87d2bada5e66b78eaa339404327a3012e5524f61.

* fix debug script

* Use HMM to seperate node from v8 in cpu

When v8 runs GC, Optimization, or something else the CPU graph spikes.
However this has very little to do with the CPU requirements of the
user's application.

As such, if the user has an I/O issue and this causes a lot of
GCScavenger runs from V8, the CPU can spike a lot which hides the I/O
issue. This uses a Hidden Markov Model to statistically separate those
two modes.