Fix race condition when reading systemInfo
This stream is being piped to `endpoint()` in two places. Previously one
of them was inside a `_transform()` method, which may be called
asynchronously at any point. It was possible for the systemInfoReader to
emit the system info _before_ this method was called, and then
TraceEventDecoder would not actually be able to access the systemInfo.

Now the pipes are set up in the same synchronous tick so both endpoints
are guaranteed to get it.

Fixes #297